feat: split FragmentationService
