+++ ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesfragment/services/FragmentationServiceImpl.java
@@ -4,51 +4,21 @@ import be.vlaanderen.informatievlaanderen.ldes.server.domain.config.LdesConfig;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.config.ViewConfig;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.entities.FragmentInfo;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.entities.LdesFragment;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.entities.LdesMember;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.repository.LdesFragmentRespository;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.repository.LdesMemberRepository;
 import org.springframework.stereotype.Component;
-
-import java.util.Optional;
 
 @Component
 public class FragmentationServiceImpl implements FragmentationService {
 
     private final LdesConfig ldesConfig;
     private final ViewConfig viewConfig;
-    private final LdesMemberRepository ldesMemberRepository;
     private final LdesFragmentRespository ldesFragmentRespository;
 
-    private final FragmentCreator fragmentCreator;
-
     public FragmentationServiceImpl(LdesConfig ldesConfig, ViewConfig viewConfig,
-                                    LdesMemberRepository ldesMemberRepository, LdesFragmentRespository ldesFragmentRespository, FragmentCreator fragmentCreator) {
+                                    LdesFragmentRespository ldesFragmentRespository) {
         this.ldesConfig = ldesConfig;
         this.viewConfig = viewConfig;
-        this.ldesMemberRepository = ldesMemberRepository;
         this.ldesFragmentRespository = ldesFragmentRespository;
-        this.fragmentCreator = fragmentCreator;
-    }
-
-    @Override
-    public LdesMember addMember(LdesMember ldesMember) {
-        ldesMember.replaceTreeMemberStatement(ldesConfig.getHostName(), ldesConfig.getCollectionName());
-        LdesFragment ldesFragment = retrieveLastFragmentOrCreateNewFragment(ldesMember);
-        ldesFragment.addMember(ldesMember.getLdesMemberId());
-        ldesFragmentRespository.saveFragment(ldesFragment);
-        return ldesMemberRepository.saveLdesMember(ldesMember);
-    }
-
-    private LdesFragment retrieveLastFragmentOrCreateNewFragment(LdesMember ldesMember) {
-        return ldesFragmentRespository.retrieveOpenFragment(ldesConfig.getCollectionName())
-                .map(fragment -> {
-                    if (fragment.getCurrentNumberOfMembers() >= viewConfig.getMemberLimit()) {
-                        return fragmentCreator.createNewFragment(Optional.of(fragment), ldesMember);
-                    } else {
-                        return fragment;
-                    }
-                })
-                .orElseGet(() -> fragmentCreator.createNewFragment(Optional.empty(), ldesMember));
     }
 
     @Override
