+++ ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesmember/services/MemberIngestService.java
@@ -1,0 +1,8 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.services;
+
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.entities.LdesMember;
+
+public interface MemberIngestService {
+
+    LdesMember addMember(LdesMember ldesMember);
+}
