+++ ldes-server-port-ingestion-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/ingestion/rest/LdesMemberIngestionController.java
@@ -1,23 +1,22 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.ingestion.rest;
 
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services.FragmentationService;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.entities.LdesMember;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.services.MemberIngestService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesmember.entities.LdesMember;
-
 @RestController
 public class LdesMemberIngestionController {
 
-    private final FragmentationService ldesReader;
+    private final MemberIngestService memberIngestService;
 
-    public LdesMemberIngestionController(final FragmentationService ldesReader) {
-        this.ldesReader = ldesReader;
+    public LdesMemberIngestionController(final MemberIngestService memberIngestService) {
+        this.memberIngestService = memberIngestService;
     }
 
     @PostMapping(value = "${ldes.collectionname}", consumes = {"application/n-quads", "application/n-triples"})
     public LdesMember ingestLdesMember(@RequestBody LdesMember ldesMember) {
-        return ldesReader.addMember(ldesMember);
+        return memberIngestService.addMember(ldesMember);
     }
 }
